Refactor: Change baseUrl & import & font within glocal-css & Layout Component
