+++ src/App.tsx
@@ -1,8 +1,13 @@
 import './styles/globals.css';
 import Router from './Router';
+import Layout from './components/layout/Layout';
 
 function App() {
-  return <Router />;
+  return (
+    <Layout>
+      <Router />
+    </Layout>
+  );
 }
 
 export default App;
