+++ src/Router.tsx
@@ -1,10 +1,8 @@
 import { BrowserRouter, Route, Routes } from 'react-router-dom';
-import Header from './header/Header';
 
 const Router = () => {
   return (
     <BrowserRouter>
-      <Header />
       <Routes>
         <Route path="/" element="" />
       </Routes>
