+++ src/styles/font.ts
@@ -2,8 +2,10 @@
 
 import { css } from 'styled-components';
 
-export const font = (size: number, weight: number, lineHeight: number = 0) => css`
+const font = (size: number, weight: number, lineHeight: number = 0) => css`
   font-size: ${size}px;
   font-weight: ${weight};
   line-height: ${lineHeight}px;
 `;
+
+export default font;
