+++ src/styles/index.ts
@@ -1,2 +1,3 @@
-export {default as flex} from './flex';
-export {default as colors} from './colors';
+export { default as flex } from './flex';
+export { default as colors } from './colors';
+export { default as font } from './font';
